pyproject.toml
[project]
name = "consistency-coach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=1.1.1",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
